Refactor: Remove null-check in Louvain application and add log message
This commit removes the unnecessary null-check in the App.java, assuming that the Louvain object will always be instantiated. This change simplifies the code and promotes better understanding of its workflow. Additionally, a logging message is added into LouvainAlgorithm.java to inform user about the file line count reading process, enhancing the transparency of the data loading operation.